python/anomaly-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.6.0
torchaudio==2.6.0
speechbrain>=0.5.16
huggingface_hub<0.20.0
numpy>=1.24.0
pydantic>=2.5.0
librosa>=0.10.1
scipy>=1.11.0
python-multipart>=0.0.6
aiohttp>=3.9.0
python-dotenv>=1.0.0

